XMLParser is now a singleton that stores a map of <filename, <document, document data> > pairs.
------------------------------

Added methods are:

setCurDocument(string) - tells the parser which file to read from. This is automatically
set after a parse, but otherwise needs to be called prior to pulling data from the parser.

removeDocument(string) - the parser attempts to remove the specified filename from its document map.
This is mostly used be a few singleton classes that only use the parser to read in their initialization data.